Fixed main layout, and got rid of clunky code
Changed the layout on the main view to a linear layout to make it more dynamic.  I was not able to get rid of the constraint layout.
I got rid othe clunky code in the setPlates fucntion and replaced it with a loop to allow better changes, and make the app smaller